% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEncoding.R
\name{convertEncoding}
\alias{convertEncoding}
\title{Convert encoding of a \code{data.frame}.}
\usage{
convertEncoding(x, from_encoding = "UTF-8", to_encoding = "ISO-8859-1")
}
\arguments{
\item{x}{Data.frame to convert}

\item{from_encoding}{Character string describing the current encoding.}

\item{to_encoding}{Character string describing the target encoding.}
}
\value{
A data.frame in the target encoding.
}
\description{
Convert encoding of a \code{data.frame}.
}
\examples{
tab_cgl <- data.frame(a = c("encodings", "nightmare"), b = c("ñ", "o"))
convertEncoding(tab_cgl)

}

