% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCgls.R
\name{queryCgls}
\alias{queryCgls}
\title{Basic table by conglomerate-date.}
\usage{
queryCgls(database, state = "all", organization = "all", cgl_id = "all",
  year_visit = 2010:2016, month_visit = 1:12)
}
\arguments{
\item{database}{Connection to an existing postgresql or sqlite database (as
returned by \code{\link[dplyr]{src_postgres}} or
\code{\link[dplyr]{src_sqlite}}.}

\item{state}{Character vector of Mexican states to include in the searh,
defaults to "all".}

\item{organization}{Character vector of organizations to include in the
search (CONAFOR, CONANP, FMCN), defaults to "all".}

\item{cgl_id}{Character vector of conglomerate ids to include in the
searh, defaults to "all".}

\item{year_visit}{Numeric vector indicating the year(s) to include in the
search, defaults to 2014:2016.}

\item{month_visit}{Numeric vector indicating the month(s) to include in the
search, defaults to 1:12.}
}
\value{
A \code{data.frame} where each line corresponds to a
  conglomerate-date, filtered by \code{state}, \code{organization},
  \code{cgl_id}, \code{year}, and \code{month}, the \code{data.frame}
  includes conglomerate name, coordinates of the conglomerate center, date of
  visit, organization, state, municipality, monitoring type, vegetation, and
  perturbated.
}
\description{
\code{queryCgls} creates basic tables by conglomerate-date.
}
\examples{

# connect to sqlite database (snmb)
database <- dplyr::src_sqlite(system.file("extdata", "snmb.sqlite",
  package = "querysnmb"))
cgl_table <- queryCgls(database)
cgl_table <- queryCgls(database, organization = "CONAFOR", year = 2014)

\dontrun{
# connect to postgreSQL database (snmb)
PASS_SNMB = Sys.getenv("PASS_SNMB")
database <- dplyr::src_postgres(dbname = "snmb", host = "dbms", user =
  "snmb", password = PASS_SNMB)
cgl_table <- queryCgls(database)
cgl_table <- queryCgls(database, organization = "CONAFOR", year = 2014)
}

}

