% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFiles.R
\name{queryFiles}
\alias{queryFiles}
\title{Paths to recordings or images}
\usage{
queryFiles(database, file_type, state = "all", organization = "all",
  cgl_id = "all", year_visit = 2010:2016, month_visit = 1:12,
  photo = TRUE, video = TRUE, ultrasonic = FALSE, original = FALSE)
}
\arguments{
\item{database}{Connection to an existing postgresql or sqlite database (as
returned by \code{\link[dplyr]{src_postgres}} or
\code{\link[dplyr]{src_sqlite}}.}

\item{file_type}{String specifing which file paths should be retrieved, value
should be "recordings" or "images" (includes images and videos from the
camera trap).}

\item{organization}{Character vector of organizations to include in the
search (CONAFOR, CONANP, FMCN), defaults to "all".}

\item{cgl_id}{Character vector of conglomerate ids to include in the
searh, defaults to "all".}

\item{year_visit}{Numeric vector indicating the year(s) to include in the
search, defaults to 2014:2016.}

\item{month_visit}{Numeric vector indicating the month(s) to include in the
search, defaults to 1:12.}

\item{photo, video, ultrasonic}{Binary variables, if
\code{file_type = "images"}, \code{photo = FALSE, video = TRUE} indicates
to only search for videos, similarly \code{video = FALSE, photo = TRUE}
excludes videos from results, defult sets both to \code{TRUE}. In case of
recordings it is only possible to download one type of file (acoustic OR
ultrasonic), defaults to download acoustic, this is
\code{ultrasonic = FALSE}.}

\item{original}{Indicates if a column with the original file name should be
included in the results.}
}
\value{
A \code{data.frame} containing the following columns:
   \code{file_type}, \code{state}, \code{institucion} (organization),
   \code{cgl_id} (conglomerate id), \code{date} (date of visit),
   \code{lat, lon} (coordinates of the conglomerate), and
   \code{archivo_nombre_original} (if original = TRUE, the original file
   name).
}
\description{
\code{queryFiles} returns paths to recordings, images or videos. For this
function to work you need to be \strong{connected to madmex folder}, this is
because the function searches for existing files in the madmex directories.
}
\examples{
# connect to sqlite database (snmb)
database <- dplyr::src_sqlite(system.file("extdata", "snmb.sqlite",
  package = "querysnmb"))
files_images_videos <- queryFiles(database, file_type = "images",
  cgl_id = "405", year_visit = 2014)
files_videos <- queryFiles(database, file_type = "images", cgl_id = "2286",
  year_visit = 2014, photo = FALSE)
files_acoustic <- queryFiles(database, file_type = "recordings",
  ultrasonic = FALSE, year_visit = 2014)
\dontrun{
# connect to postgreSQL database (snmb)
PASS_SNMB = Sys.getenv("PASS_SNMB")
database <- dplyr::src_postgres(dbname = "snmb", host = "dbms", user =
  "snmb", password = PASS_SNMB)
files_acoustic_name <- queryFiles(database, file_type = "recordings",
  ultrasonic = FALSE, year_visit = 2014, original = TRUE)
}

}

