% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querysnmb.R
\docType{package}
\name{querysnmb}
\alias{querysnmb}
\alias{querysnmb-package}
\title{querysnmb: A package for querying the SNMB database.}
\description{
The querysnmb package provides three categories of important functions:
foo, bar and baz.
}
\section{Querysnmb functions}{

The foo functions ...
}

