% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querySpecies.R
\name{querySpecies}
\alias{queryCamera}
\alias{queryFootprints}
\alias{queryInvaders}
\alias{querySpecies}
\alias{querySpecimens}
\title{Query tables by species.}
\usage{
queryInvaders(database, state = "all", organization = "all",
  cgl_id = "all", year_visit = 2010:2016, month_visit = 1:12,
  noms = "all")

queryFootprints(database, state = "all", organization = "all",
  cgl_id = "all", year_visit = 2010:2016, month_visit = 1:12,
  noms = "all")

querySpecimens(database, state = "all", organization = "all",
  cgl_id = "all", year_visit = 2010:2016, month_visit = 1:12,
  noms = "all")

queryCamera(database, state = "all", organization = "all", cgl_id = "all",
  year_visit = 2010:2016, month_visit = 1:12, noms = "all")
}
\arguments{
\item{database}{Connection to an existing postgresql or sqlite database (as
returned by \code{\link[dplyr]{src_postgres}} or
\code{\link[dplyr]{src_sqlite}}.}

\item{state}{Character vector of Mexican states to include in the searh,
defaults to "all".}

\item{organization}{Character vector of organizations to include in the
search (CONAFOR, CONANP, FMCN), defaults to "all".}

\item{cgl_id}{Character vector of conglomerate ids to include in the
searh, defaults to "all".}

\item{year_visit}{Numeric vector indicating the year(s) to include in the
search, defaults to 2014:2016.}

\item{month_visit}{Numeric vector indicating the month(s) to include in the
search, defaults to 1:12.}

\item{noms}{String indicating the species names to look for (ex:
"vaca|equus")}
}
\value{
A \code{data.frame} where each line corresponds to a
conglomerate-date, filtered by \code{state}, \code{organization},
\code{cgl_id}, \code{year}, and \code{month}, the \code{data.frame}
includes conglomerate name, coordinates of the conglomerate center,
date of visit, organization, state, municipality, monitoring type,
vegetation, perturbated, species common and scientific name, ....
}
\description{
Search for species records in Invaders, Footprints or Specimens tables, the
results are indexed by conglomerate-date and may be filtered by state,
organization, date, or conglomerate ids.
}
\examples{

# connect to sqlite database (snmb)
database <- dplyr::src_sqlite(system.file("extdata", "snmb.sqlite",
  package = "querysnmb"))
invaders <- queryInvaders(database)
specimens <- querySpecimens(database, state = "Chiapas")

\dontrun{
connect to database (snmb)
PASS_SNMB = Sys.getenv("PASS_SNMB")
database <- dplyr::src_postgres(dbname = "snmb", host = "dbms", user =
  "snmb", password = PASS_SNMB)
invaders <- queryInvaders(database, noms = "vaca|bos|perro",
  year_visit = 2014:2015, organization = c("CONANP", "FMCN"))
footprints <- queryFootprints(database, noms = "vaca|bos|equus|caballo")
}
}

