% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querysnmb.R
\docType{package}
\name{querysnmb}
\alias{querysnmb}
\alias{querysnmb-package}
\title{querysnmb: A package for querying the SNMB database.}
\description{
The querysnmb package provides functions to query the SNMB database in an
amicable manner, this is, the user does not need to know the database schema
or repeateadly perform annoying joins to have useful databases.
}
\section{Querysnmb functions}{

The functions use the dplyr package to query the database, the advantage of
using them is that the query functions perform the necessary joins to have a
useful table of information, and provide several variables to filter the
results by.
}

